Integrated abilities with permissions
-Added methods to adjust and update a user's abilities based on their level
-Added more methods in User to facilitate adding, removing, and retrieving abilities
-Added new commands to set a user and input's level, update user abilities, and toggle a command's enabled state
-Moved all commands into a Commands folder

+++ TRBot/TRBot.Commands/BaseCommand.cs
@@ -33,7 +33,7 @@ namespace TRBot.Commands
     {
         public bool Enabled = true;
         public bool DisplayInHelp = true;
-        public int Level = 0;
+        public long Level = 0;
         public string ValueStr = string.Empty;
 
         protected CommandHandler CmdHandler = null;
